+++ src/pages/_app.js
@@ -1,5 +1,6 @@
 import Head from "next/head";
 import "@/styles/app.css";
+import Navbar from "./components/Navbar";
 
 export default function App({ Component, pageProps }) {
   return (
@@ -7,6 +8,7 @@ export default function App({ Component, pageProps }) {
       <Head>
         <title>SCRC 2024</title>
       </Head>
+      <Navbar />
       <Component {...pageProps} />
     </>
   );
