+++ src/pages/components/Background.jsx
