+++ src/pages/faq/index.jsx
