--- src/styles/about.css
@@ -1,158 +1,0 @@
-#about {
-  display: flex;
-}
-
-.plane-tv-frame {
-  background: white;
-  border: 0.5em solid black;
-  border-radius: 1.8em;
-  width: 100%;
-  height: 90dvh;
-  margin: auto;
-  position: relative;
-}
-
-.plane-tv-nav-container {
-  width: 100%;
-  height: 6%;
-  display: flex;
-  justify-content: space-evenly;
-  align-items: center;
-  border-top: 0.2em solid black;
-  position: absolute;
-  bottom: 0;
-}
-
-.plane-tv-nav-button {
-  width: calc(100% / 3);
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.plane-tv-nav-button.left-button {
-  border-radius: 0 0 0 1.2em;
-}
-
-.plane-tv-nav-button.middle-button {
-  border-right: 0.5px solid black;
-  border-left: 0.5px solid black;
-}
-
-.plane-tv-nav-button.right-button {
-  border-radius: 0 0 1.2em 0;
-}
-
-.plane-tv-nav-button:hover {
-  background-color: var(--blue5);
-  color: white;
-}
-
-.plane-tv-nav-button:active {
-  background-color: var(--blue6);
-  color: white;
-}
-
-.tab-selected {
-  background-color: var(--blue6);
-  color: white;
-}
-
-.tab-selected.plane-tv-nav-button:hover {
-  background-color: var(--blue6);
-  color: white;
-}
-
-.about-tab-container {
-  width: 100%;
-  height: 94%;
-  padding: 1em;
-}
-
-.about-tab-container h3 {
-  font-size: 1.1em;
-  font-weight: bold;
-  text-decoration: underline;
-  text-align: center;
-}
-
-.about-tab-container p {
-  margin-block-end: 1em;
-  text-align: center;
-}
-
-#trip-details-tab {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-around;
-}
-
-.tamu-map {
-  width: 100%;
-  height: 70%;
-  color: white;
-  background: black;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-@media (min-width: 481px) {
-  /* portrait e-readers (Nook/Kindle), smaller tablets @ 600 or @ 640 wide. */
-  .plane-tv-frame {
-    border-radius: 1.6em;
-  }
-
-  .plane-tv-nav-button.left-button {
-    border-radius: 0 0 0 1em;
-  }
-
-  .plane-tv-nav-button.right-button {
-    border-radius: 0 0 1em 0;
-  }
-}
-@media (min-width: 641px) {
-  /* portrait tablets, portrait iPad, landscape e-readers, landscape 800x480 or 854x480 phones */
-  .plane-tv-frame {
-    height: 70dvh;
-    border-radius: 1.3em;
-  }
-
-  .plane-tv-nav-container {
-    position: unset;
-    border-top: none;
-    border-bottom: 0.2em solid black;
-  }
-  .plane-tv-nav-button.left-button {
-    border-radius: 0.6em 0 0 0;
-  }
-
-  .plane-tv-nav-button.right-button {
-    border-radius: 0 0.6em 0 0;
-  }
-}
-@media (min-width: 961px) {
-  /* tablet, landscape iPad, lo-res laptops ands desktops */
-}
-@media (min-width: 1025px) {
-  /* big landscape tablets, laptops, and desktops */
-}
-@media (min-width: 1281px) {
-  /* hi-res laptops and desktops */
-  .plane-tv-frame {
-    border-radius: 1em;
-  }
-
-  .plane-tv-nav-button.left-button {
-    border-radius: 0.4em 0 0 0;
-  }
-
-  .plane-tv-nav-button.right-button {
-    border-radius: 0 0.4em 0 0;
-  }
-}
-@media (min-width: 1921px) {
-  /* hi-res desktops */
-}
