--- src/styles/event-page.css
@@ -1,79 +1,0 @@
-.event-container {
-  background: white;
-  color: black;
-  flex-direction: column;
-  align-items: center;
-  position: absolute;
-  top: 0;
-  height: 100%;
-  width: 100%;
-  animation: fade-in 1s forwards;
-}
-
-.event-container > h2 {
-  flex: 1;
-  font-size: 2em;
-  text-decoration: underline;
-  font-weight: 700;
-}
-
-.event-description-container {
-  flex: 10;
-  width: 100%;
-  padding: 0.5em;
-}
-
-.event-close-container {
-  flex: 1;
-  width: 100%;
-  padding: 0.5em;
-}
-
-#tba {
-}
-
-#registration {
-}
-
-#opening-ceremony {
-}
-
-#workshop-1 {
-}
-
-#workshop-2 {
-}
-
-#lunch-break {
-}
-
-#networking-session {
-}
-
-#workshop-3 {
-}
-
-#closing-ceremony {
-}
-
-#career-fair {
-}
-
-@media (min-width: 481px) {
-  /* portrait e-readers (Nook/Kindle), smaller tablets @ 600 or @ 640 wide. */
-}
-@media (min-width: 641px) {
-  /* portrait tablets, portrait iPad, landscape e-readers, landscape 800x480 or 854x480 phones */
-}
-@media (min-width: 961px) {
-  /* tablet, landscape iPad, lo-res laptops ands desktops */
-}
-@media (min-width: 1025px) {
-  /* big landscape tablets, laptops, and desktops */
-}
-@media (min-width: 1281px) {
-  /* hi-res laptops and desktops */
-}
-@media (min-width: 1921px) {
-  /* hi-res desktops */
-}
