--- src/styles/guide.css
@@ -1,260 +1,0 @@
-#key-dates {
-  display: flex;
-}
-
-#schedule-content-container {
-  margin-top: var(--header-size);
-}
-
-.schedule-container {
-  width: 100%;
-  height: 90dvh;
-  margin: auto;
-  position: relative;
-}
-
-.schedule-board {
-  background: white;
-  width: 100%;
-  height: 100%;
-}
-
-.schedule-board .schedule-divider {
-  transition: transform 1s ease;
-  transform-style: preserve-3d;
-  width: 100%;
-  position: relative;
-}
-
-.schedule-board .header-container {
-  width: 100%;
-  height: 10%;
-}
-
-#schedule-header-logo {
-  margin-right: 1em;
-}
-
-h2.event-title {
-  flex: 2;
-  font-size: 1.08em;
-  text-align: left;
-}
-
-h2.long-title {
-  display: none;
-  flex: 4;
-  font-size: 1.35em;
-  text-align: right;
-}
-
-h2.short-title {
-  flex: 4;
-  font-size: 1.35em;
-  text-align: right;
-}
-
-.schedule-board > .label {
-  background: var(--blue6);
-  color: white;
-  width: 100%;
-  height: 4.5%;
-  font-size: 0.72em;
-}
-
-.label > div > h2,
-.front-divider > h2 {
-  line-height: 100%;
-  margin: auto;
-}
-
-h2.departure-label {
-  flex: 2;
-  text-align: left;
-}
-
-h2.gate-label {
-  flex: 1;
-  text-align: center;
-}
-
-h2.start-time-label {
-  text-align: center;
-  flex: 2;
-}
-
-h2.status-label {
-  flex: 1;
-  display: none;
-  text-align: center;
-}
-
-.schedule-board > .time-divider {
-  color: white;
-  width: 100%;
-  height: 7%;
-  font-size: 1.17em;
-}
-
-.schedule-board .event-divider {
-  width: 100%;
-  height: 6%;
-  font-size: 0.81em;
-}
-
-.schedule-board .event-divider > button:hover {
-  background: var(--blue3);
-  color: white;
-}
-
-.schedule-board .event-divider > button:active {
-  background: var(--blue4);
-  color: white;
-}
-
-.schedule-board .transition-divider {
-  color: white;
-  width: 100%;
-  height: 3.5%;
-  font-size: 0.72em;
-}
-
-.blue5-divider {
-  background: var(--blue5);
-}
-
-.schedule-board .pad-divider {
-  padding: 0 0.5em;
-}
-
-.front-divider,
-.back-divider {
-  backface-visibility: hidden;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: absolute;
-}
-
-.front-divider {
-  border-bottom: 0.5px solid black;
-}
-
-.header-container .front-divider {
-  border-top: 0.5px solid black;
-  padding: 0.5em;
-}
-
-.back-divider {
-  background: white;
-  transform: rotateX(180deg);
-}
-
-.flip-divider {
-  transform: rotateX(180deg);
-}
-
-@keyframes slide-container-in-from-top {
-  from {
-    z-index: 1;
-    transform: translateY(-125dvh);
-  }
-  to {
-    z-index: 1;
-    transform: translateY(0dvh);
-  }
-}
-
-@keyframes slide-container-out-to-top {
-  0% {
-    z-index: 1;
-    transform: translateY(0dvh);
-  }
-  99% {
-    z-index: 1;
-  }
-  100% {
-    z-index: -1;
-    transform: translateY(-125dvh);
-  }
-}
-
-@media (min-width: 481px) {
-  /* portrait e-readers (Nook/Kindle), smaller tablets @ 600 or @ 640 wide. */
-}
-@media (min-width: 641px) {
-  /* portrait tablets, portrait iPad, landscape e-readers, landscape 800x480 or 854x480 phones */
-  h2.long-title {
-    display: block;
-  }
-
-  h2.short-title {
-    display: none;
-  }
-
-  h2.status-label {
-    display: block;
-  }
-}
-@media (min-width: 961px) {
-  /* tablet, landscape iPad, lo-res laptops ands desktops */
-}
-@media (min-width: 1025px) {
-  /* big landscape tablets, laptops, and desktops */
-}
-@media (min-width: 1281px) {
-  /* hi-res laptops and desktops */
-  h2.event-title {
-    flex: 2;
-    font-size: 1.2em;
-    text-align: left;
-  }
-
-  h2.long-title {
-    flex: 4;
-    font-size: 1.5em;
-    text-align: right;
-  }
-
-  h2.short-title {
-    flex: 4;
-    font-size: 1.5em;
-    text-align: right;
-  }
-
-  .schedule-board > .label {
-    background: var(--blue6);
-    color: white;
-    width: 100%;
-    height: 4.5%;
-    font-size: 0.8em;
-  }
-
-  .schedule-board > .time-divider {
-    color: white;
-    width: 100%;
-    height: 7%;
-    font-size: 1.3em;
-  }
-
-  .schedule-board .event-divider {
-    width: 100%;
-    height: 6%;
-    font-size: 0.9em;
-  }
-
-  .schedule-board .transition-divider {
-    color: white;
-    width: 100%;
-    height: 3.5%;
-    font-size: 0.8em;
-  }
-}
-@media (min-width: 1921px) {
-  /* hi-res desktops */
-  .schedule-container {
-    height: 70dvh;
-  }
-}
