--- src/styles/landing-page.css
@@ -1,169 +1,0 @@
-/* smartphones, iPhone, portrait 480x320 phones */
-section#landing-page {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-evenly;
-  user-select: none;
-}
-
-.splash-title {
-  display: flex;
-  flex-direction: column;
-  min-width: 260px;
-  width: 50%;
-  margin: 0 auto;
-}
-
-.splash-title h1 {
-  font-size: 3.375em;
-  line-height: 1em;
-}
-
-#sase-title {
-  width: 75%;
-  aspect-ratio: 1/1;
-}
-
-#rc-title {
-  margin-block-end: 0.3em;
-}
-
-#at-title {
-  margin-inline-start: 0.5em;
-}
-
-.registration-link:hover {
-  background-color: var(--blue2);
-}
-
-.key-info-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 75%;
-  margin: 0 auto;
-}
-
-.registration-link {
-  font-size: 1.5em;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 12em;
-  height: 2.5em;
-  background-color: rgb(255, 255, 255);
-  border-radius: 1em;
-  transition: border 0.2s ease-in;
-  transition: background-color 0.2s ease-in;
-}
-
-.landing-date {
-  position: relative;
-  left: 50%;
-  top: 40%;
-  transform: translate(-14%, -50%);
-}
-
-.landing-date > h2 {
-  font-size: 1.5em;
-}
-.banner {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-}
-
-.date-banner {
-  background: url("/banner-plane/banner.svg") no-repeat;
-  background-size: contain;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding-left: 25px;
-  margin: 2em 0em 2em 75px;
-  width: 264px;
-  height: 75px;
-  font-size: 20px;
-  animation: banner-fly-in-left 3.125s linear 1 forwards;
-}
-
-.right-facing-plane {
-  background: url("/banner-plane/plane_without_string.svg") no-repeat;
-  background-size: contain;
-  width: 100px;
-  height: 35px;
-  animation: plane-fly-in-out-left 5s linear 1 forwards;
-}
-
-@media (min-width: 481px) {
-  /* portrait e-readers (Nook/Kindle), smaller tablets @ 600 or @ 640 wide. */
-}
-@media (min-width: 641px) {
-  /* portrait tablets, portrait iPad, landscape e-readers, landscape 800x480 or 854x480 phones */
-  section#landing-page {
-    display: flex;
-    flex-direction: row;
-    justify-content: inherit;
-    user-select: none;
-  }
-
-  .splash-title {
-    width: fit-content;
-    margin: auto 0;
-  }
-
-  .key-info-container {
-    margin: auto;
-  }
-
-  .banner {
-    display: inherit;
-    animation: balloon-float 8s ease-in-out infinite;
-    margin: 3em 0 0 0;
-  }
-
-  .date-banner {
-    background: url("/airship/blimp.svg") no-repeat;
-    background-size: contain;
-    width: 300px;
-    height: 120px;
-    padding-left: 0;
-    margin: 0;
-    animation: none;
-  }
-
-  .right-facing-plane {
-    display: none;
-    width: 0;
-    height: 0;
-  }
-
-  .banner-label {
-    position: absolute;
-    top: 38px;
-    left: 90px;
-  }
-}
-@media (min-width: 961px) {
-  /* tablet, landscape iPad, lo-res laptops ands desktops */
-}
-@media (min-width: 1025px) {
-  /* big landscape tablets, laptops, and desktops */
-}
-@media (min-width: 1281px) {
-  /* hi-res laptops and desktops */
-}
-@media (min-width: 1921px) {
-  /* hi-res desktops */
-  .sase-title {
-    font-size: 9em;
-  }
-
-  .sc-title {
-    font-size: 5em;
-  }
-
-  .rc-title {
-    font-size: 5em;
-  }
-}
