+++ src/styles/modal.css
@@ -22,7 +22,7 @@
   width: 100%;
   height: 100%;
   background: rgba(0, 0, 0, 0.5);
-  z-index: 1;
+  z-index: 10000;
 }
 
 .hide-modal {
