+++ src/styles/navbar.css
@@ -4,42 +4,50 @@
   top: 0;
   width: 100%;
   min-height: var(--header-size);
-  background-color: rgba(255, 255, 255, 0.9);
-  border-bottom: 1px outset var(--blue1);
-  backdrop-filter: blur(3px);
   user-select: none;
-  transition: height 0.5s;
+  transition: height 0.5s ease;
+  transition: background-color 0.2s ease;
+  z-index: 1000;
+}
+
+.top-navbar {
+  background: transparent;
+  color: black;
+}
+
+.scrolled-navbar {
+  background-color: rgba(255, 255, 255, 0.9);
+  backdrop-filter: blur(3px);
+  color: black;
 }
 
 .navbar-box {
   display: flex;
   align-items: center;
-  justify-content: space-between;
-  width: 100%;
   height: var(--header-size);
   margin-left: auto;
   margin-right: auto;
 }
 
 #navbar-logo {
-  display: none;
+  margin-right: 0.4rem;
+  width: 85%;
+  height: 85%;
 }
 
 #header-title {
-  float: left;
-  color: black;
-  text-transform: uppercase;
   display: flex;
   align-items: center;
   height: 100%;
   text-decoration: none;
+  text-transform: uppercase;
   transition: color background-image 0.3s ease-in;
-  padding: 0.5rem;
 }
 
 #header-title > h1 {
   font-size: 1.5em;
   font-weight: 700;
+  margin: auto 0;
 }
 
 #header-title:hover {
@@ -48,27 +56,44 @@
 
 .horizontal-navbar {
   display: none;
-  flex-direction: row;
   align-items: center;
+  margin-left: auto;
   height: 100%;
-  float: right;
 }
 
 .dropdown-navbar {
+  position: fixed;
+  top: 0;
+  transform: translateY(-100%);
   display: flex;
   flex-direction: column;
-  padding: 0.5rem 0.5rem 0.75rem 0.5rem;
+  padding: 0rem 0.5rem 0.5rem 0.5rem;
+  width: 100%;
+  border-bottom: 1px outset var(--blue1);
+  background: white;
+  z-index: 999;
 }
 
 .dropdown-navbar-icon {
-  display: inline-flex;
+  display: flex;
   align-items: center;
-  padding: 0.5em 1.1em 0.5em 0.5em;
-  border-radius: 0.375rem;
-  align-items: center;
-  justify-content: right;
+  padding: 0 0.5em;
   height: 100%;
-  width: 100%;
+  width: calc(20px + 1em);
+}
+
+.menu-open-icon {
+  background: url("/icons/burger-black-icon.svg") no-repeat;
+  background-size: contain;
+  height: 20px;
+  width: 20px;
+}
+
+.menu-close-icon {
+  background: url("/icons/close-black-icon.svg") no-repeat;
+  background-size: contain;
+  height: 20px;
+  width: 20px;
 }
 
 .horizontal-navbar .navbar-link {
@@ -81,12 +106,18 @@
 
 .horizontal-navbar .navbar-link:hover {
   color: var(--blue5);
+  box-shadow: 0px 0px 0px 1px black inset;
+}
+
+.horizontal-navbar li {
+  display: inline;
 }
 
 .dropdown-navbar .navbar-link {
   display: block;
   padding: 0.5rem;
-  border-radius: 0.5rem;
+  border-radius: 0.5rem 0.5rem 0 0;
+  border-bottom: 1px solid black;
 }
 
 .dropdown-navbar .navbar-link:hover {
@@ -105,6 +136,42 @@
   z-index: 3;
 }
 
+#navbar-backdrop {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: rgba(0, 0, 0, 0.5);
+  z-index: 998;
+}
+
+.open-nav {
+  animation: slide-down-nav 0.5s ease-out forwards;
+}
+
+.close-nav {
+  animation: slide-up-nav 0.8s ease-in forwards;
+}
+
+@keyframes slide-down-nav {
+  from {
+    transform: translateY(-100%);
+  }
+  to {
+    transform: translateY(0);
+  }
+}
+
+@keyframes slide-up-nav {
+  from {
+    transform: translateY(0);
+  }
+  to {
+    transform: translateY(-100%);
+  }
+}
+
 @media (min-width: 481px) {
   /* portrait e-readers (Nook/Kindle), smaller tablets @ 600 or @ 640 wide. */
 }
@@ -112,7 +179,10 @@
   /* portrait tablets, portrait iPad, landscape e-readers, landscape 800x480 or 854x480 phones */
   #navbar-logo {
     margin-right: 1rem;
-    display: initial;
+  }
+
+  #header-title {
+    margin-left: 0.5rem;
   }
 
   .horizontal-navbar {
